% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart.R
\name{bar_chart}
\alias{bar_chart}
\title{bar_chart}
\usage{
bar_chart(
  label,
  width = "100\%",
  marginLeft = "8px",
  height = "16px",
  fill = "#fc5185",
  background = NULL
)
}
\description{
https://glin.github.io/reactable/articles/cookbook/cookbook.html
}
